strands-examples/strands-langfuse-demo/pyproject.toml
[project]
name = "strands-langfuse-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents[otel]==1.6.0",
    "streamlit==1.49.1",
]
